requirements.txt
# RunPod SDK
runpod==1.4.0

# Transformers e dependências de IA
transformers==4.44.0
torch==2.3.0+cu121 --index-url https://download.pytorch.org/whl/cu121
accelerate
safetensors
Pillow
requests
